pyproject.toml
[project]
name = "ayoub-ai-assistant"
version = "1.0.0"
description = "Ayoub — JARVIS-style AI assistant with ReAct, voice, MCP tools, and Ollama"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # ── Core ──────────────────────────────────────────────────────────────────
    "python-dotenv",
    "rich",
    "colorama",

    # ── LLM providers ─────────────────────────────────────────────────────────
    "google-generativeai>=0.7.0",
    "langchain-google-genai",
    "langchain>=0.3.0",
    "langchain-core",
    "langchain-experimental",
    "langchain-openai",
    "langchain-groq",
    "openai",
    "groq",
    "ollama",

    # ── MCP server ────────────────────────────────────────────────────────────
    "fastmcp",
    "httpx",

    # ── Voice / LiveKit (optional) ────────────────────────────────────────────
    "livekit-agents[openai,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",

    # ── Web / scraping ────────────────────────────────────────────────────────
    "beautifulsoup4",
    "lxml",
    "requests",

    # ── File tools ────────────────────────────────────────────────────────────
    "PyPDF2",

    # ── Vision / screen ───────────────────────────────────────────────────────
    "pillow",
    "opencv-python",

    # ── Image generation ──────────────────────────────────────────────────────
    "gradio_client",
    "matplotlib",
]

[project.scripts]
# uv generates ayoub.exe on Windows, ayoub script on Linux — no bash needed
ayoub        = "ayoub.cli:main"
ayoub-server = "ayoub.mcp_server.server:main"
ayoub-voice  = "ayoub.voice.agent:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ayoub"]
